src/requirements.txt
streamlit
pandas
PyPDF2
requests
python-dotenv
sounddevice
numpy
matplotlib
pyaudio
wave
SpeechRecognition
